refactor: scielomanager/notifications.py #1027
Fixes: #1027

Separação do código genérico (scielomanager/notifications.py @ Message)
do código particular (articletrack/notifications.py @ CheckinMessage, TicketMessage)
e (editorialmanager/notifications.py @ IssueBoardMessage, BoardMemberMessage)

Os tests de cada Clase *Message estão dentro da pasta de tests de cada app.
