+++ scielomanager/editorialmanager/tests/tests_notifications.py
@@ -4,19 +4,11 @@ from django.core import mail
 from django.conf import settings
 from django.test import TestCase
 from django.test.utils import override_settings
-
 from django_factory_boy import auth
 
 from journalmanager.tests import modelfactories
-
-from scielomanager import notifications
-
+from editorialmanager import notifications
 from . import modelfactories as editorial_modelfactories
-
-def generate_subject(action, subject=''):
-    subject_prefix = settings.EMAIL_SUBJECT_PREFIX
-    subject_suffix = notifications.EMAIL_DATA_BY_ACTION[action]['subject_sufix']
-    return ' '.join([subject_prefix, subject, subject_suffix])
 
 
 class IssueBoardMessageTests(TestCase):
@@ -30,6 +22,15 @@ class IssueBoardMessageTests(TestCase):
         self.journal = modelfactories.JournalFactory.create(editor=self.editor)
         self.issue = modelfactories.IssueFactory(journal=self.journal)
         self.expected_recipients = [self.editor, ]
+        self.expected_subject_sufix_by_action = {
+            'issue_add_no_replicated_board': "Issue Board can't be replicated",
+            'issue_add_replicated_board': "Issue has a new replicated board",
+        }
+
+    def _make_subject(self, action, subject=''):
+        subject_prefix = settings.EMAIL_SUBJECT_PREFIX
+        subject_suffix = self.expected_subject_sufix_by_action[action]
+        return ' '.join([subject_prefix, subject, subject_suffix])
 
     def tearDown(self):
         """
@@ -41,7 +42,7 @@ class IssueBoardMessageTests(TestCase):
         email_count = 0
         for action in self.ACTIONS:
             # with
-            expected_subject = generate_subject(action=action, subject='')
+            expected_subject = self._make_subject(action)
             email_count += 1
             # when
             result = notifications.issue_board_replica(issue=self.issue, action=action)
@@ -53,6 +54,7 @@ class IssueBoardMessageTests(TestCase):
             self.assertEqual(len(self.expected_recipients), len(new_email.to))
             for recipient in self.expected_recipients:
                 self.assertIn(recipient, new_email.to)
+
 
 class BoardMembersMessageTests(TestCase):
     ACTIONS =  [
@@ -80,13 +82,23 @@ class BoardMembersMessageTests(TestCase):
         self.member = editorial_modelfactories.EditorialMemberFactory.create(board=self.board)
 
         self.expected_recipients = [self.librarian1, self.librarian2, ]
+        self.expected_subject_sufix_by_action = {
+            'board_add_member': "Member of the journal board, was added",
+            'board_edit_member': "Member of the journal board, was edited",
+            'board_delete_member': "Member of the journal board, was deleted",
+        }
+
+    def _make_subject(self, action, subject=''):
+        subject_prefix = settings.EMAIL_SUBJECT_PREFIX
+        subject_suffix = self.expected_subject_sufix_by_action[action]
+        return ' '.join([subject_prefix, subject, subject_suffix])
 
     @override_settings(CELERY_EAGER_PROPAGATES_EXCEPTIONS=True, CELERY_ALWAYS_EAGER=True, BROKER_BACKEND='memory')
     def test_each_action(self):
         email_count = 0
         for action in self.ACTIONS:
             # with
-            expected_subject = generate_subject(action=action, subject='')
+            expected_subject = self._make_subject(action)
             message = ''
             email_count += 1
             # when
